Trie: reduce cell size - by using int16 for prefix lengths (#17134)
`536` -> `488` bytes

+++ execution/commitment/commitmentdb/commitment_context.go
@@ -556,7 +556,7 @@ func (sdc *TrieContext) Storage(plainKey []byte) (u *commitment.Update, err erro
 	}
 	u = &commitment.Update{
 		Flags:      commitment.DeleteUpdate,
-		StorageLen: len(enc),
+		StorageLen: int8(len(enc)),
 	}
 
 	if u.StorageLen > 0 {
